Split up clientContext and subscriptionContext
limit re-rendering of client when possible
